Removes the singleton stuff from these instances and simplifies lots of tests.
Why singleton are not needed:

Making this instances singletons leads to subtle bugs in queries and
complicates the code just to avoid such complications:

An instance may be involved in a query::

   query = these(parent for parent in this('parent')
   	   		    if parent.age > 40)

   query2 = these(parent for parent in this('parent')
   	    		    if parent.age < 30)

If this('parent') were to return a singleton then the second query
could have overwritten the previous binding from the first query; it
took a bit of hackerish to avoid this: Create a context when iterating
over these instances and then skip the singleton-making code in such a
context; so this('parent') would create separate instances.

But since These instance are likely to be used always inside queries
the singleton stuff would not make improvement on the general case. So
it is best just to remove it.