Renames IQuery to IGeneratorToken.
This is done because the instances are actually equivalent
to the syntactical portion we can identify with a "<- DB" in UnQL.

This is called generator in languages like Haskell and Erlang. However
since the word "generator" does have a meaning the name GeneratorToken
may change in the future if we figure out a better one.