Introduces a `tokens` in QueryPart, explained in DesignDecisionsRegressionTests.
Basically it's need in the case of query expressions like:

 (parent.title + parent.name, child.name + child.nick)
 for parent in this('parent')
 for child in parent.children
 if (parent.age > 32) & (child.age < 5) | contains(child.toys, 'laptop')

The filters were wrongly generated cause we only notified the *primary*
part involved: this way `this('parent').children` got the notifications
for `child.age < 5` and `contains(child.toys, 'laptop')` but never
was notified of the relation of both expressions.

So `these` extracted those expressions as filters of the entire query which
is wrong. The interpretation of the entire query be:

- Filter by (parents with age greater than 32 and child with
  age lesser than 5) OR by child that have 'laptop' in his toys collection.

- AND the child has less than 5

- AND the child has 'laptop' in his toys collection.

Notice that last to filters are invalid.

The solution is to notify to all involved tokens.