No changes. Explains the escaping particles.
It seems that there's bug in xoutil.context regarding as to which context
it returns when searching by a Interface. There are now failing tests in 
xoutil's repository.

Until this is fixed is best just to use a "normal" key for contexts in these
and there create a stack of bubbles.