Fixes context-related bug by implementing bubbles stack locally.
As noted previously (commit id: c1cad651c737e91ab7be6aeb73cfdce63bdfdf31)
there's bug with storing data in contexts where those are looked for
an interface key.

And since we may remove looking for interfaces in a future version of
xoutil (mainly because of this bug), we prefer just to implement
the stack of bubbles here using a thread-local variable.

TODO: Prove that `these` is thread-safe.