Fixes stability of filters.
If `set` is used, then the order of filters is not guarranted to be stable (as
in the query expression). Furthermore if a programmer creates the same filter
twice is not the job of the query language to remove the duplicate. This is
left to an optimization phase in translators.