Removes QueryPart. This implies a big commit with several changes in design.
The removal was provoked because QueryParts won't actually help in the case of
the subquery like::

    these(parent
          for parent in this('parent')
          if any_(child for child in parent.children if child.age < 6))

In this case the previous mechanism of letting the first argument (a QueryPart
usually) to be who emits the built expression fails, because the argument to
``any_`` was a generator object.

So expressions would have to emit themselves if any bubble was around. That
fact removed the need of a QueryPart.

But several edge cases arouse:

- Cross-bubbling of `parent.children` (part, token) duality. Since the part is
  emitted to the upper bubble, when the token is emitted it does not realizes
  it sibling part was emitted and cannot squash it.