Fixes bug: thesefy impedes super.
The problem was that thesefy actually modified the class hierarchy, by creating
a returning a new class that inherited from the original one. So super(Target,
self) actually retrieved the hidden original class instead of its base, and all
methods were actually in that hidden original class.

Problem diagram::

        Before thesefy

        +-----------------+
        |  Foobar class   |
        +-----------------+
        |  attrs          |
	|  .....          |
        +-----------------+

        After thesefy

        +------------------------+
        |  hidden Foobar class   |
        +------------------------+
        |  attrs                 |
	|  .....                 |
        +------------------------+
             ^
             |
             |  inherits
        +-----------------+
        |  Foobar class   |
        +-----------------+
        +-----------------+

The solution is to recreate the Foobar from the original's attributes. This is
already what `xoutil.decorator.compat.metaclass` does, we just use it here.