Fix issue with Empty collections.
Empty() was not an instance of Empty and also when trying to recognize
the `xs` in a Cons to be Empty the was a type mismatch.

`is_collection` is too restrictive, the `itertools.chain` instance
returned by the split in _head is not a collection.  So we now use
`is_iterable`.

Notice, however, that an empty iterable will not be considered an
instance of Empty.  We guard against that peeking the iterables inside
`_head` and directly returning `Empty()` when the iterable yields
nothing.