Remove code left behind on the instance/type duality of Cons and Foldr.
Cons and Foldr were first programmed to be replaced by partial
instances::

   Cons = Cons(Undefined, Undefined)

so you'll always import an instance... But that was deemed too
complicated and would harm future instances recognition (``isinstance``)
that we now deem needed when considering the composition of expressions.

We still need to craft how "resolving" (calling) is appropriate in this
scenario.