Save locals and globals.
Translators need to access the references to locals/globals where the
query was defined.  Specifically translators we'll need to manage the
internal '.0' name given for the first iterable in a comprehension.

When we construct the QST for the comprehension

   ``(x for x in this)``

the ``this`` reference is internalized and the name seen in the QST is
'.0', so translators need to compare the value of '.0' to see if this is
the Universe instance.

Furthermore, for a query like

   ``((x, CONSTANT) for x in this)``

the translators have to access the value of ``CONSTANT`` to properly
build results.