Fix the naive plan for monadic plans.
I added a test that demonstrate the failure.  First I explain the test
and then I proceed to tackle the issue.

In the test I have three plans:

``plan``
  An explicitly built plan with all monadic operators fully expanded.

``plan1``
  A translated plan that iterates over a Cons.

``plan2``
  A translated plan that iterates over this.  Currently this fails to
  execute.

The translated plans should be equivalent in structure to the explicit
one.  The difference is the collection being traversed.

In the translated plans the collection is a ``qst.Name('.0',
qst.Load())`` whereas in the explicit plan a Cons is directly passed.
In the ``plan1`` resolves to a Cons, whereas in the ``plan2`` is
resolves to a QueryObject.

Notice I don't test for equality of results but that all plans fail or
pass consistently.  So this test may help spot future errors.  The
following analysis is, thus, applicable to the current failure.

The problem must be in what '.0' is returning when iterating over
``this``.

So far this is the procedure:

1. 'translate' creates a plan and compiles it.

2. When calling the plan the compiled plan is evaluated on a single
   dictionary that chains the query frame's locals and globals, and the
   plan's operators (map, join, unit, and zero).

   While chaining, it iterates over all locals and replaces any
   instances of the ``this`` value for an instance of a
   PythonObjectsCollection.

   While chaining, it iterates over all globals and replaces any
   instances of the ``this`` value for an instance of a
   PythonObjectsCollections.

   This means that the '.0' will be resolved to an instance of
   PythonObjectsCollections.

3. The PythonObjectsCollection is actually the argument of a Map, so it
   will be iterated over.

   And when using a fully monadic plan a Cons-like construction is
   expected.

   The problem was that we were passing the entire iterable to the Cons
   and it expects a head and a tail.

   Notice however that Cons is recursive and the PythonObjectsCollection
   is too vast to and a RuntimeError (maximum recursion depth exceeded)
   is now the problem.