Fix walking of CALL_FUNCTION.
As stated in 81ba6118c03545cb1ccf4286cda7037972d70f34, the star-argument now
comes before all keywords.  Thus we need to update the walker to match this
stack order.

Additionally we make explicit (in the grammar) what's an start-argument
and (double-star) kw-argument.  This makes it easier to debug the parser.