Avoid Unset as absent attribute marker.
Up to xoutil 2.0.2, `Unset == 0` because Unset is a boolean which inherits
from `int`.

In our case, comparing a `qst.Num(0, qst.Load())` with anything else would be
True, because we would try to compare the attribute `n=0` with Unset.

This impacted in the search for a sentinel in the stack of the walker.