Speed up pops from the stack.
Simple experiments show up that `pop_n` is up to 10x faster.  With a random
stack of 308 items, the previous implementation showed:

>>> %timeit pop_n(list(stack), n)
13.9 µs ± 52.5 ns per loop (mean ± std. dev. of 7 runs, 100_000 loops each)

This implementation:

>>> %timeit pop_n(list(stack), n)
1.68 µs ± 30.8 ns per loop (mean ± std. dev. of 7 runs, 1_000_000 loops each)

Standard deviation is high.  But notice that the later run an order of
magnitude higher than the first.

For `n` bigger than stack the speed up is even better, because the loop is
completely avoided.

`pop_until_timeout` shows less improvements.  I think the `indexlast`
algorithm can be improved.