Speed up the last-index algorithm.
Experimental results:

Our algorithm

    >>> %timeit lastindex(stack, 22)
    3.77 µs ± 23.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

The previous one

    >>> %timeit lastindex(stack, 22)
    10.3 µs ± 41.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

The previous algorithm starts at the start (that's how `index()` works) and
proceeds one item at a time until it finds the item or the reaches the end (or
stop).  Then it looks again for the same item after the point where the
instance was found, if no other instance is found, return the previously found
index.  This last procedure is repeated until no other instance is found.

This algorithm loops the entire list always.

The new algorithm halves the list and calls `index()` over the second half.
If there's an instance there we ignore completely the first half.  Then it
"recurses" over the selected half.