Really speed up the lastindex algorithm.
It hit me that the halving and other stuff was just overestimating the
slowness of Python code.

The 'obvious' version is still faster than the previous attempts.  I've left
the obvious in the tests.

Obvious version stats:

>> %timeit lastindex(stack, 21)
1.71 µs ± 99.9 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

However the fastest version (the one left in the commit) is unbeatable:

>>> %timeit lastindex(stack, 21)
856 ns ± 3.61 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

I thinks this is faster because:

- It's almost all implemented in C.

- We avoid picking items from lists, since the comparison is made by `index`
  itself.